enable sharding

+++ index.php
@@ -1,21 +1,15 @@
 <?php include("layout/header.php"); ?>
 <h1>Welcome to MongoAdmin for sharding</h1>
 <hr>
-<?php if(!empty($_GET['error'])){ ?>
-    <div class="alert alert-block alert-danger"><?php echo $_GET['error']; ?></div>
-<?php } ?>
-<?php if(!empty($_GET['success'])){ ?>
-    <div class="alert alert-block alert-success"><?php echo $_GET['success']; ?></div>
-<?php } ?>
 <?php $results = $mongoDB->getShards(); ?>
 <h4>Add a shard</h4>
     <form action="sharding.php" method="post" role="form" class="form-inline">
         <div class="form-group">
-            <label for="shardIp">IP</label>
+            <label for="shardIp">IP:</label>
             <input type="text" class="form-control" id="shardIp" name="shardIP" placeholder="127.0.0.1">
         </div>
         <div class="form-group">
-            <label for="shardName">Name</label>
+            <label for="shardName">Name:</label>
             <input type="text" class="form-control" id="shardName" name="shardName" placeholder="myShard">
         </div>
         <button type="submit" class="btn btn-success" name="addShard" value="ok">Add</button>
@@ -32,7 +26,7 @@
         <tr>
             <td><?php echo $shard['_id']; ?></td>
             <td><?php echo $shard['host']; ?></td>
-            <td><a href="sharding.php?delete=<?php echo $shard['_id']; ?>">x</a></td>
+            <td><a onclick="return confirm('Are you sure?');" href="sharding.php?delete=<?php echo $shard['_id']; ?>">x</a></td>
         </tr>
         <?php } ?>
     </table>
